requirements_pyredmine.txt
-r requirements.txt
python-redmine
